minor fix
git-svn-id: https://forge.indepnet.net/svn/glpi/trunk@11370 35612d4c-9d10-0410-837c-bf9dd1be60d7

+++ front/group.form.php
@@ -71,7 +71,7 @@ if (isset($_POST["add"])) {
    glpi_header($_SERVER['HTTP_REFERER']);
 
 } else if (isset($_POST["deleteuser"])) {
-   if (count($_POST["item"])) {
+   if (isset($_POST["item"]) && count($_POST["item"])) {
       foreach ($_POST["item"] as $key => $val) {
          if ($groupuser->can($key,'w')) {
             $groupuser->delete(array('id'=>$key));
